[INTERNAL] Log message when using EOF Node.js versions
For Node.js < 14.16, we can't show the update-notifier anymore as it
requires a higher Node.js version.
See: https://github.com/SAP/ui5-cli/pull/533

As those versions are end-of-life and not supported / maintained by the
Node.js team, we log a message. It also lets users know that there
might be a new UI5 CLI version available.