[INTERNAL] New dynamic import mock for NodeJS >= 20.6.xx (#663)
Fixes: https://github.com/SAP/ui5-tooling/issues/891

Node's module loading has changed since 20.6.xx. As `esmock` is unable
to properly mock dynamic imports within CommonJS modules, we need to use
another alternative for this test. `testdouble` seems to work pretty
well for that case, but for NodeJS versions below 20.6.xx it needs a
`--loader` flag and we already use that for `esmock`.

The solution would be to use the old mocking for older NodeJs versions
and `testdouble` for the ones after 20.6.xx (no flag required here):
https://github.com/testdouble/testdouble.js/blob/main/docs/7-replacing-dependencies.md#how-module-replacement-works-for-es-modules-using-import